tag-instance-on-create/requirements.txt
functions-framework==3.2.0
google-cloud-compute==1.5.2